settings jwt 8 hours
Probably not a good idea, but given the always open nature of todo
app’s I feel like this is necessary. At least until I find a better way
to handle this, I’ve noted a few but none seem great. And I’m not keen
on using Auth0 for this, though it seems great.